requirements_dev.txt
pip==19.0.3
twine==1.13.0
wheel==0.33.1
PyYAML==5.1
colorlog==3.1.0
numpy==1.13.1
pandas